requirements.txt
# sandbox had python 3.10.6
# pydantic==2.8.2
gradio==4.44.1
huggingface_hub>=0.29.3
together==1.2.0
python-dotenv~=1.0.1
